requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
crispy-bootstrap5==0.6
Django==4.0.3
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djlint==0.7.6
idna==3.3
importlib-metadata==4.11.3
mccabe==0.7.0
pathspec==0.9.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==1.7.1
pylama==8.3.8
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
snowballstemmer==2.2.0
sqlparse==0.4.2
tomlkit==0.9.2
tqdm==4.63.0
tzdata==2022.1
urllib3==1.26.9
zipp==3.7.0
